fix error

+++ search_algorithms/0-linear.c
@@ -10,7 +10,7 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	int cont;
+	size_t cont;
 	cont = 0;
 
 	if (array == NULL)
